requirements.txt
langchain
langchain_openai
streamlit
pandas
openpyxl
python-dotenv
